+++ core/src/lib_helpers.rs
@@ -11,8 +11,8 @@ use crate::model::repo::RepoSource;
 use crate::model::state::Config;
 use crate::pure_functions::files;
 use crate::repo::account_repo;
-use crate::repo::file_repo;
 use crate::service::drawing_service::SupportedImageFormats;
+use crate::service::file_service;
 use crate::service::{drawing_service, file_encryption_service};
 use crate::CoreError;
 
@@ -20,7 +20,7 @@ pub fn get_children_helper(
     config: &Config,
     id: Uuid,
 ) -> Result<Vec<ClientFileMetadata>, CoreError> {
-    let files = file_repo::get_all_not_deleted_metadata(config, RepoSource::Local)?;
+    let files = file_service::get_all_not_deleted_metadata(config, RepoSource::Local)?;
     let files = files::filter_not_deleted(&files);
     let children = files::find_children(&files, id);
     children
@@ -33,7 +33,7 @@ pub fn get_and_get_children_recursively_helper(
     config: &Config,
     id: Uuid,
 ) -> Result<Vec<FileMetadata>, CoreError> {
-    let files = file_repo::get_all_not_deleted_metadata(config, RepoSource::Local)?;
+    let files = file_service::get_all_not_deleted_metadata(config, RepoSource::Local)?;
     let files = files::filter_not_deleted(&files);
     let file_and_descendants = files::find_with_descendants(&files, id)?;
 
@@ -56,14 +56,14 @@ pub fn get_and_get_children_recursively_helper(
 }
 
 pub fn delete_file_helper(config: &Config, id: Uuid) -> Result<(), CoreError> {
-    let files = file_repo::get_all_not_deleted_metadata(config, RepoSource::Local)?;
+    let files = file_service::get_all_not_deleted_metadata(config, RepoSource::Local)?;
     let file = files::apply_delete(&files, id)?;
-    file_repo::insert_metadatum(config, RepoSource::Local, &file)
+    file_service::insert_metadatum(config, RepoSource::Local, &file)
 }
 
 pub fn read_document_helper(config: &Config, id: Uuid) -> Result<DecryptedDocument, CoreError> {
-    let all_metadata = file_repo::get_all_metadata(config, RepoSource::Local)?;
-    file_repo::get_not_deleted_document(config, RepoSource::Local, &all_metadata, id)
+    let all_metadata = file_service::get_all_metadata(config, RepoSource::Local)?;
+    file_service::get_not_deleted_document(config, RepoSource::Local, &all_metadata, id)
 }
 
 pub fn save_document_to_disk_helper(
@@ -71,30 +71,30 @@ pub fn save_document_to_disk_helper(
     id: Uuid,
     location: String,
 ) -> Result<(), CoreError> {
-    let all_metadata = file_repo::get_all_metadata(config, RepoSource::Local)?;
+    let all_metadata = file_service::get_all_metadata(config, RepoSource::Local)?;
     let document =
-        file_repo::get_not_deleted_document(config, RepoSource::Local, &all_metadata, id)?;
+        file_service::get_not_deleted_document(config, RepoSource::Local, &all_metadata, id)?;
     files::save_document_to_disk(&document, location)
 }
 
 pub fn rename_file_helper(config: &Config, id: Uuid, new_name: &str) -> Result<(), CoreError> {
-    let files = file_repo::get_all_not_deleted_metadata(config, RepoSource::Local)?;
+    let files = file_service::get_all_not_deleted_metadata(config, RepoSource::Local)?;
     let files = files::filter_not_deleted(&files);
     let file = files::apply_rename(&files, id, new_name)?;
-    file_repo::insert_metadatum(config, RepoSource::Local, &file)
+    file_service::insert_metadatum(config, RepoSource::Local, &file)
 }
 
 pub fn move_file_helper(config: &Config, id: Uuid, new_parent: Uuid) -> Result<(), CoreError> {
-    let files = file_repo::get_all_not_deleted_metadata(config, RepoSource::Local)?;
+    let files = file_service::get_all_not_deleted_metadata(config, RepoSource::Local)?;
     let files = files::filter_not_deleted(&files);
     let file = files::apply_move(&files, id, new_parent)?;
-    file_repo::insert_metadatum(config, RepoSource::Local, &file)
+    file_service::insert_metadatum(config, RepoSource::Local, &file)
 }
 
 pub fn get_drawing_helper(config: &Config, id: Uuid) -> Result<Drawing, CoreError> {
-    let all_metadata = file_repo::get_all_metadata(config, RepoSource::Local)?;
+    let all_metadata = file_service::get_all_metadata(config, RepoSource::Local)?;
     let drawing_bytes =
-        file_repo::get_not_deleted_document(config, RepoSource::Local, &all_metadata, id)?;
+        file_service::get_not_deleted_document(config, RepoSource::Local, &all_metadata, id)?;
     drawing_service::parse_drawing(&drawing_bytes)
 }
 
@@ -104,8 +104,8 @@ pub fn save_drawing_helper(
     drawing_bytes: &[u8],
 ) -> Result<(), CoreError> {
     drawing_service::parse_drawing(drawing_bytes)?; // validate drawing
-    let metadata = file_repo::get_not_deleted_metadata(config, RepoSource::Local, id)?;
-    file_repo::insert_document(config, RepoSource::Local, &metadata, drawing_bytes)
+    let metadata = file_service::get_not_deleted_metadata(config, RepoSource::Local, id)?;
+    file_service::insert_document(config, RepoSource::Local, &metadata, drawing_bytes)
 }
 
 pub fn export_drawing_helper(
@@ -114,9 +114,9 @@ pub fn export_drawing_helper(
     format: SupportedImageFormats,
     render_theme: Option<HashMap<ColorAlias, ColorRGB>>,
 ) -> Result<Vec<u8>, CoreError> {
-    let all_metadata = file_repo::get_all_metadata(config, RepoSource::Local)?;
+    let all_metadata = file_service::get_all_metadata(config, RepoSource::Local)?;
     let drawing_bytes =
-        file_repo::get_not_deleted_document(config, RepoSource::Local, &all_metadata, id)?;
+        file_service::get_not_deleted_document(config, RepoSource::Local, &all_metadata, id)?;
     drawing_service::export_drawing(&drawing_bytes, format, render_theme)
 }
 
@@ -127,9 +127,9 @@ pub fn export_drawing_to_disk_helper(
     render_theme: Option<HashMap<ColorAlias, ColorRGB>>,
     location: String,
 ) -> Result<(), CoreError> {
-    let all_metadata = file_repo::get_all_metadata(config, RepoSource::Local)?;
+    let all_metadata = file_service::get_all_metadata(config, RepoSource::Local)?;
     let drawing_bytes =
-        file_repo::get_not_deleted_document(config, RepoSource::Local, &all_metadata, id)?;
+        file_service::get_not_deleted_document(config, RepoSource::Local, &all_metadata, id)?;
     let exported_drawing_bytes =
         drawing_service::export_drawing(&drawing_bytes, format, render_theme)?;
     files::save_document_to_disk(&exported_drawing_bytes, location)
