+++ core/src/repo/mod.rs
@@ -2,7 +2,6 @@ pub mod account_repo;
 pub mod db_version_repo;
 pub mod digest_repo;
 pub mod document_repo;
-pub mod file_repo;
 pub mod last_updated_repo;
 pub mod local_storage;
 pub mod metadata_repo;
