+++ core/src/service/path_service.rs
@@ -6,8 +6,9 @@ use lockbook_models::file_metadata::FileType::{Document, Folder};
 use crate::model::repo::RepoSource;
 use crate::model::state::Config;
 use crate::pure_functions::files;
-use crate::repo::{account_repo, file_repo};
+use crate::repo::account_repo;
 
+use crate::service::file_service;
 use crate::CoreError;
 
 pub fn create_at_path(
@@ -23,7 +24,7 @@ pub fn create_at_path(
 
     let is_folder = path_and_name.ends_with('/');
 
-    let mut files = file_repo::get_all_not_deleted_metadata(config, RepoSource::Local)?;
+    let mut files = file_service::get_all_not_deleted_metadata(config, RepoSource::Local)?;
     let mut current = files::find_root(&files)?;
     let root_id = current.id;
     let account = account_repo::get(config)?;
@@ -67,7 +68,7 @@ pub fn create_at_path(
 
         current = files::apply_create(&files, file_type, current.id, next_name, &account.username)?;
         files.push(current.clone());
-        file_repo::insert_metadatum(config, RepoSource::Local, &current)?;
+        file_service::insert_metadatum(config, RepoSource::Local, &current)?;
     }
 
     Ok(current)
@@ -76,7 +77,7 @@ pub fn create_at_path(
 pub fn get_by_path(config: &Config, path: &str) -> Result<DecryptedFileMetadata, CoreError> {
     let paths = split_path(path);
 
-    let files = file_repo::get_all_not_deleted_metadata(config, RepoSource::Local)?;
+    let files = file_service::get_all_not_deleted_metadata(config, RepoSource::Local)?;
     let mut current = files::find_root(&files)?;
 
     for (i, &value) in paths.iter().enumerate() {
@@ -123,7 +124,7 @@ pub fn filter_from_str(input: &str) -> Result<Option<Filter>, CoreError> {
 }
 
 pub fn get_all_paths(config: &Config, filter: Option<Filter>) -> Result<Vec<String>, CoreError> {
-    let files = file_repo::get_all_not_deleted_metadata(config, RepoSource::Local)?;
+    let files = file_service::get_all_not_deleted_metadata(config, RepoSource::Local)?;
 
     let mut filtered_files = files.clone();
 
@@ -158,7 +159,7 @@ pub fn get_all_paths(config: &Config, filter: Option<Filter>) -> Result<Vec<Stri
 }
 
 pub fn get_path_by_id(config: &Config, id: Uuid) -> Result<String, CoreError> {
-    let files = file_repo::get_all_not_deleted_metadata(config, RepoSource::Local)?;
+    let files = file_service::get_all_not_deleted_metadata(config, RepoSource::Local)?;
     let mut current_metadata = files::find(&files, id)?;
     let mut path = String::from("");
 
@@ -193,9 +194,9 @@ mod unit_tests {
 
     use crate::model::repo::RepoSource;
     use crate::model::state::temp_config;
-    use crate::repo::{account_repo, file_repo};
+    use crate::repo::account_repo;
     use crate::service::path_service::Filter;
-    use crate::service::{path_service, test_utils};
+    use crate::service::{file_service, path_service, test_utils};
     use crate::{files, CoreError};
 
     #[test]
@@ -205,7 +206,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let doc = path_service::create_at_path(config, &format!("{}/document", &account.username))
             .unwrap();
@@ -220,7 +221,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let folder =
             path_service::create_at_path(config, &format!("{}/folder/", &account.username))
@@ -236,7 +237,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let folder =
             path_service::create_at_path(config, &format!("{}/folder/", &account.username))
@@ -256,7 +257,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let document =
             path_service::create_at_path(config, &format!("{}/folder/document", &account.username))
@@ -275,7 +276,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let document = path_service::create_at_path(
             config,
@@ -300,7 +301,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let result =
             path_service::create_at_path(config, &format!("{}//document", &account.username));
@@ -315,7 +316,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let result = path_service::create_at_path(
             config,
@@ -332,7 +333,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         path_service::create_at_path(config, &format!("{}/folder/document", &account.username))
             .unwrap();
@@ -349,7 +350,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         path_service::create_at_path(config, &format!("{}/not-folder", &account.username)).unwrap();
         let result = path_service::create_at_path(
@@ -367,7 +368,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let created_document =
             path_service::create_at_path(config, &format!("{}/document", &account.username))
@@ -385,7 +386,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let created_folder =
             path_service::create_at_path(config, &format!("{}/folder/", &account.username))
@@ -403,7 +404,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let created_document =
             path_service::create_at_path(config, &format!("{}/folder/document", &account.username))
@@ -422,7 +423,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let document =
             path_service::create_at_path(config, &format!("{}/document", &account.username))
@@ -439,7 +440,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let folder =
             path_service::create_at_path(config, &format!("{}/folder/", &account.username))
@@ -456,7 +457,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         let document =
             path_service::create_at_path(config, &format!("{}/folder/document", &account.username))
@@ -476,7 +477,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         path_service::create_at_path(
             config,
@@ -515,7 +516,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         path_service::create_at_path(
             config,
@@ -542,7 +543,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         path_service::create_at_path(
             config,
@@ -578,7 +579,7 @@ mod unit_tests {
         let root = files::create_root(&account.username);
 
         account_repo::insert(config, &account).unwrap();
-        file_repo::insert_metadatum(config, RepoSource::Local, &root).unwrap();
+        file_service::insert_metadatum(config, RepoSource::Local, &root).unwrap();
 
         path_service::create_at_path(
             config,
