+++ core/src/service/sync_service.rs
@@ -18,8 +18,8 @@ use crate::model::repo::RepoState;
 use crate::model::state::Config;
 use crate::pure_functions::files;
 use crate::repo::account_repo;
-use crate::repo::{file_repo, last_updated_repo};
-use crate::service::{api_service, file_encryption_service};
+use crate::repo::last_updated_repo;
+use crate::service::{api_service, file_encryption_service, file_service};
 use crate::CoreError;
 
 use super::file_compression_service;
@@ -40,7 +40,7 @@ pub fn calculate_work(config: &Config) -> Result<WorkCalculated, CoreError> {
     info!("Calculating Work");
 
     let account = account_repo::get(config)?;
-    let base_metadata = file_repo::get_all_metadata(config, RepoSource::Base)?;
+    let base_metadata = file_service::get_all_metadata(config, RepoSource::Base)?;
     let base_max_metadata_version = base_metadata
         .iter()
         .map(|f| f.metadata_version)
@@ -65,7 +65,7 @@ fn calculate_work_from_updates(
     mut last_sync: u64,
 ) -> Result<WorkCalculated, CoreError> {
     let mut work_units: Vec<WorkUnit> = vec![];
-    let (all_metadata, _) = file_repo::get_all_metadata_with_encrypted_changes(
+    let (all_metadata, _) = file_service::get_all_metadata_with_encrypted_changes(
         config,
         RepoSource::Local,
         server_updates,
@@ -80,7 +80,7 @@ fn calculate_work_from_updates(
             last_sync = metadata.metadata_version;
         }
 
-        match file_repo::maybe_get_metadata(config, RepoSource::Local, metadata.id)? {
+        match file_service::maybe_get_metadata(config, RepoSource::Local, metadata.id)? {
             None => {
                 if !metadata.deleted {
                     // no work for files we don't have that have been deleted
@@ -105,12 +105,12 @@ fn calculate_work_from_updates(
             .cmp(&f2.get_metadata().metadata_version)
     });
 
-    for file_diff in file_repo::get_all_metadata_changes(config)? {
-        let metadata = file_repo::get_metadata(config, RepoSource::Local, file_diff.id)?;
+    for file_diff in file_service::get_all_metadata_changes(config)? {
+        let metadata = file_service::get_metadata(config, RepoSource::Local, file_diff.id)?;
         work_units.push(WorkUnit::LocalChange { metadata });
     }
-    for doc_id in file_repo::get_all_with_document_changes(config)? {
-        let metadata = file_repo::get_metadata(config, RepoSource::Local, doc_id)?;
+    for doc_id in file_service::get_all_with_document_changes(config)? {
+        let metadata = file_service::get_metadata(config, RepoSource::Local, doc_id)?;
         work_units.push(WorkUnit::LocalChange { metadata });
     }
     debug!("Work Calculated: {:#?}", work_units);
@@ -419,7 +419,7 @@ fn get_resolved_document(
         .iter()
         .find(|&f| f.clone().local().id == remote_metadatum.id);
     let maybe_document_state = if let Some(metadata_state) = maybe_metadata_state {
-        file_repo::maybe_get_document_state(config, metadata_state)?
+        file_service::maybe_get_document_state(config, metadata_state)?
     } else {
         None
     };
@@ -481,7 +481,7 @@ fn pull<F>(
 where
     F: FnMut(SyncProgressOperation),
 {
-    let base_metadata = file_repo::get_all_metadata(config, RepoSource::Base)?;
+    let base_metadata = file_service::get_all_metadata(config, RepoSource::Base)?;
     let base_max_metadata_version = base_metadata
         .iter()
         .map(|f| f.metadata_version)
@@ -501,13 +501,13 @@ where
     .map_err(CoreError::from)?
     .file_metadata;
 
-    let local_metadata = file_repo::get_all_metadata(config, RepoSource::Local)?;
-    let (remote_metadata, remote_orphans) = file_repo::get_all_metadata_with_encrypted_changes(
+    let local_metadata = file_service::get_all_metadata(config, RepoSource::Local)?;
+    let (remote_metadata, remote_orphans) = file_service::get_all_metadata_with_encrypted_changes(
         config,
         RepoSource::Base,
         &remote_metadata_changes,
     )?;
-    let all_metadata_state = file_repo::get_all_metadata_state(config)?;
+    let all_metadata_state = file_service::get_all_metadata_state(config)?;
 
     let num_documents_to_pull = remote_metadata_changes
         .iter()
@@ -638,7 +638,7 @@ where
     // resolve cycles
     for self_descendant in files::get_invalid_cycles(&local_metadata, &local_metadata_updates)? {
         if let Some(RepoState::Modified { mut local, base }) =
-            file_repo::maybe_get_metadata_state(config, self_descendant)?
+            file_service::maybe_get_metadata_state(config, self_descendant)?
         {
             if let Some(existing_update) =
                 files::maybe_find_mut(&mut local_metadata_updates, self_descendant)
@@ -646,20 +646,20 @@ where
                 existing_update.parent = base.parent;
             }
             local.parent = base.parent;
-            file_repo::insert_metadatum(config, RepoSource::Local, &local)?;
+            file_service::insert_metadatum(config, RepoSource::Local, &local)?;
         }
     }
 
     // update base
-    file_repo::insert_metadata(config, RepoSource::Base, &base_metadata_updates)?;
+    file_service::insert_metadata(config, RepoSource::Base, &base_metadata_updates)?;
     for (metadata, document_update) in base_document_updates {
-        file_repo::insert_document(config, RepoSource::Base, &metadata, &document_update)?;
+        file_service::insert_document(config, RepoSource::Base, &metadata, &document_update)?;
     }
 
     // update local
-    file_repo::insert_metadata(config, RepoSource::Local, &local_metadata_updates)?;
+    file_service::insert_metadata(config, RepoSource::Local, &local_metadata_updates)?;
     for (metadata, document_update) in local_document_updates {
-        file_repo::insert_document(config, RepoSource::Local, &metadata, &document_update)?;
+        file_service::insert_document(config, RepoSource::Local, &metadata, &document_update)?;
     }
 
     Ok(())
@@ -679,7 +679,7 @@ where
     ));
 
     // update remote to local (metadata)
-    let metadata_changes = file_repo::get_all_metadata_changes(config)?;
+    let metadata_changes = file_service::get_all_metadata_changes(config)?;
     if !metadata_changes.is_empty() {
         api_service::request(
             account,
@@ -691,7 +691,7 @@ where
     }
 
     // update base to local
-    file_repo::promote_metadata(config)?;
+    file_service::promote_metadata(config)?;
 
     Ok(())
 }
@@ -705,9 +705,9 @@ fn push_documents<F>(
 where
     F: FnMut(SyncProgressOperation),
 {
-    for id in file_repo::get_all_with_document_changes(config)? {
-        let mut local_metadata = file_repo::get_metadata(config, RepoSource::Local, id)?;
-        let local_content = file_repo::get_document(config, RepoSource::Local, &local_metadata)?;
+    for id in file_service::get_all_with_document_changes(config)? {
+        let mut local_metadata = file_service::get_metadata(config, RepoSource::Local, id)?;
+        let local_content = file_service::get_document(config, RepoSource::Local, &local_metadata)?;
         let encrypted_content = file_encryption_service::encrypt_document(
             &file_compression_service::compress(&local_content)?,
             &local_metadata,
@@ -730,14 +730,14 @@ where
         .new_content_version;
 
         // save content version change
-        let mut base_metadata = file_repo::get_metadata(config, RepoSource::Base, id)?;
+        let mut base_metadata = file_service::get_metadata(config, RepoSource::Base, id)?;
         base_metadata.content_version = local_metadata.content_version;
-        file_repo::insert_metadatum(config, RepoSource::Local, &local_metadata)?;
-        file_repo::insert_metadatum(config, RepoSource::Base, &base_metadata)?;
+        file_service::insert_metadatum(config, RepoSource::Local, &local_metadata)?;
+        file_service::insert_metadatum(config, RepoSource::Base, &base_metadata)?;
     }
 
     // update base to local
-    file_repo::promote_documents(config)?;
+    file_service::promote_documents(config)?;
 
     Ok(())
 }
@@ -751,7 +751,7 @@ pub fn sync(
     config: &Config,
     maybe_update_sync_progress: Option<Box<dyn Fn(SyncProgress)>>,
 ) -> Result<(), CoreError> {
-    let mut sync_progress_total = 4 + file_repo::get_all_with_document_changes(config)?.len(); // 3 metadata pulls + 1 metadata push + doc pushes
+    let mut sync_progress_total = 4 + file_service::get_all_with_document_changes(config)?.len(); // 3 metadata pulls + 1 metadata push + doc pushes
     let mut sync_progress = 0;
     let mut update_sync_progress = |op: SyncProgressOperation| match op {
         SyncProgressOperation::IncrementTotalWork(inc) => sync_progress_total += inc,
@@ -769,12 +769,12 @@ pub fn sync(
 
     let account = &account_repo::get(config)?;
     pull(config, account, &mut update_sync_progress)?;
-    file_repo::prune_deleted(config)?;
+    file_service::prune_deleted(config)?;
     push_metadata(config, account, &mut update_sync_progress)?;
     pull(config, account, &mut update_sync_progress)?;
     push_documents(config, account, &mut update_sync_progress)?;
     pull(config, account, &mut update_sync_progress)?;
-    file_repo::prune_deleted(config)?;
+    file_service::prune_deleted(config)?;
     last_updated_repo::set(config, get_time().0)?;
     Ok(())
 }
